pyproject.toml
[tool.poetry]
name = "rna_lm"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
gensim = "^4.2.0"
Logbook = "^1.5.3"
scikit-learn = "^1.1.1"
biopython = "^1.79"
lightgbm = "^3.3.2"
biotite = "^0.34.0"
pandas = "^1.0.1"
torch = "^1.12.1"
torchvision = "^0.13.1"
attrdict = "^2.0.1"
ipykernel = "^6.15.3"
pyarrow = "^9.0.0"
nltk = "^3.7"
fastparquet = "^0.8.3"
numpy = "^1.23.3"
scipy = "1.8.1"
PyYAML = "^6.0"
shap = "^0.41.0"
seaborn = "^0.12.0"
databricks-cli = "^0.17.3"
tokenizers = "^0.13.3"
evaluate = "^0.4.0"
transformers = "^4.29.0"
accelerate = "^0.19.0"
modAL-python = "^0.4.2.1"
skorch = "^0.9.0"
dvc = "^2.51.0"
dvc-s3 = "^2.21.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pylint = "^2.17.4"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
